Move the class GardenPlantingItemViewModel to own file
