+++ app/src/main/java/com/google/samples/apps/sunflower/adapters/GardenPlantingAdapter.kt
@@ -16,10 +16,8 @@
 
 package com.google.samples.apps.sunflower.adapters
 
-import android.arch.lifecycle.ViewModel
 import android.content.Context
 import android.databinding.DataBindingUtil
-import android.databinding.ObservableField
 import android.support.v7.recyclerview.extensions.ListAdapter
 import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
@@ -27,8 +25,7 @@ import android.view.ViewGroup
 import com.google.samples.apps.sunflower.R
 import com.google.samples.apps.sunflower.data.PlantAndGardenPlantings
 import com.google.samples.apps.sunflower.databinding.ListItemGardenPlantingBinding
-import java.text.SimpleDateFormat
-import java.util.*
+import com.google.samples.apps.sunflower.viewmodels.GardenPlantingItemViewModel
 
 class GardenPlantingAdapter(
     val context: Context
@@ -57,40 +54,13 @@ class GardenPlantingAdapter(
 
         fun bind(plantings: PlantAndGardenPlantings) {
             with(binding) {
-                vm = GardenPlantingItemViewModel(itemView.context, plantings)
+                vm = GardenPlantingItemViewModel(
+                    itemView.context,
+                    plantings
+                )
                 executePendingBindings()
             }
         }
     }
 
-    class GardenPlantingItemViewModel(context: Context, plantings: PlantAndGardenPlantings) :
-        ViewModel() {
-
-        private val plant = checkNotNull(plantings.plant)
-        private val gardenPlanting = plantings.gardenPlantings[0]
-
-        private val dateFormat by lazy { SimpleDateFormat("MMM d, yyyy", Locale.US) }
-        private val plantDateString by lazy { dateFormat.format(gardenPlanting.plantDate.time) }
-        private val waterDateString by lazy { dateFormat.format(gardenPlanting.lastWateringDate.time) }
-        private val wateringPrefix by lazy {
-            context.getString(R.string.watering_next_prefix, waterDateString)
-        }
-        private val wateringSuffix by lazy {
-            context.resources.getQuantityString(
-                R.plurals.watering_next_suffix,
-                plant.wateringInterval, plant.wateringInterval
-            )
-        }
-
-        val imageUrl = ObservableField<String>(plant.imageUrl)
-
-        val plantDate = ObservableField<String>(
-            context.getString(
-                R.string.planted_date, plant.name,
-                plantDateString
-            )
-        )
-
-        val waterDate = ObservableField<String>("$wateringPrefix - $wateringSuffix")
-    }
 }
