+++ app/src/main/res/layout/list_item_garden_planting.xml
@@ -22,7 +22,7 @@
 
         <variable
             name="vm"
-            type="com.google.samples.apps.sunflower.adapters.GardenPlantingAdapter.GardenPlantingItemViewModel"/>
+            type="com.google.samples.apps.sunflower.viewmodels.GardenPlantingItemViewModel"/>
     </data>
 
     <android.support.constraint.ConstraintLayout
